Intent: Centralize window validation for history coherence Operators involved: [Coherence, Self-organization] Affected invariants: [#1, #7] Key changes: - Move validate_window into tnfr.validators and export it centrally - Update glyph_history to lazily resolve the validator without cycles - Point observers and tests to the new validate_window location Expected risks/dissonances: Guarded against circular imports via cached resolver Metrics: C(t), Si, νf, phase unchanged by refactor Equivalence map: tnfr.glyph_history.validate_window -> tnfr.validators.validate_window
