+++ src/tnfr/validators.py
@@ -2,11 +2,11 @@
 
 from __future__ import annotations
 
+import numbers
 import sys
 
 from .constants import get_aliases, get_param
 from .alias import get_attr
-from .glyph_history import last_glyph
 from .sense import sigma_vector_from_graph
 from .helpers.numeric import within_range
 from .constants_glyphs import GLYPHS_CANONICAL_SET
@@ -14,7 +14,23 @@ from .constants_glyphs import GLYPHS_CANONICAL_SET
 ALIAS_EPI = get_aliases("EPI")
 ALIAS_VF = get_aliases("VF")
 
-__all__ = ("run_validators",)
+__all__ = ("validate_window", "run_validators")
+
+
+def validate_window(window: int, *, positive: bool = False) -> int:
+    """Validate ``window`` as an ``int`` and return it.
+
+    Non-integer values raise :class:`TypeError`. When ``positive`` is ``True``
+    the value must be strictly greater than zero; otherwise it may be zero.
+    Negative values always raise :class:`ValueError`.
+    """
+
+    if isinstance(window, bool) or not isinstance(window, numbers.Integral):
+        raise TypeError("'window' must be an integer")
+    if window < 0 or (positive and window == 0):
+        kind = "positive" if positive else "non-negative"
+        raise ValueError(f"'window'={window} must be {kind}")
+    return int(window)
 
 
 def _require_attr(data, alias, node, name):
@@ -52,6 +68,8 @@ def _check_glyph(g, n):
 
 def run_validators(G) -> None:
     """Run all invariant validators on ``G`` with a single node pass."""
+    from .glyph_history import last_glyph
+
     epi_min = float(get_param(G, "EPI_MIN"))
     epi_max = float(get_param(G, "EPI_MAX"))
     vf_min = float(get_param(G, "VF_MIN"))
